requirements.txt
psutil==5.2.2
mock==2.0.0
cffi==1.10.0
pytz==2017.2
Flask==0.12.2
Pygments==2.2.0
pytest==3.2.1
nose==1.3.7
matplotlib==2.1.2
gspread==3.1.0
Unirest==1.1.7
lxml==3.8.0
setuptools==39.2.0
MarkupSafe==1.0
pycrypto==2.6.1
Image==1.5.27
python_Levenshtein==0.12.0
Numeric==24.2
Pillow==5.4.1
Pyro4==4.75
annoy==1.15.1
autograd==1.2
ca_certs_locater==1.0
Django==2.1.7
elementtree==1.2.6-20050316
epdb==0.15.1
fasteners==0.14.1
genapi==0.0.8
protobuf==3.7.0
jsonpickle==1.1
keras==2.2.4
keyring==18.0.0
morfessor==2.0.4
moto==1.3.7
mpmath==1.1.0
mtrand==0.1
ndg_httpsclient==0.5.1
numarray==1.5.1
ordereddict==1.1
paramiko==2.4.2
PasteScript==3.1.0
pattern==3.6
pickle5==0.0.6
plotly==3.7.0
pyemd==0.5.1
pygtrie==2.3
requests_kerberos==0.12.0
responses==0.10.6
scikits.talkbox==0.2.5
simplejson==3.16.0
scikit_learn==0.20.3
sparsesvd==0.2.2
spline==1.12
streams==0.1
sympy==1.3
testfixtures==6.6.1
theano==1.0.4
visdom==0.1.8.8
webapp2==2.5.2
WebOb==1.8.5
PyYAML==5.1
